Move Detacher interface to the component
